Move Managers and Subscribers to correct folders.
