+++ src/Aquifer.Jobs/Managers/SyncApiRequestStatsToStorageTableManager.cs
@@ -7,7 +7,7 @@ using Microsoft.Azure.Functions.Worker;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 
-namespace Aquifer.Jobs;
+namespace Aquifer.Jobs.Managers;
 
 public class SyncApiRequestStatsToStorageTableManager(
     IAquiferApiManagementClient _apiManagementClient,
